+++ cinecore/modelos/IngressoInteira.cs
@@ -8,19 +8,19 @@ namespace cinecore.modelos
     public class IngressoInteira : Ingresso
     {
         [Required(ErrorMessage = "O preço é obrigatório")]
-        [Range(0, float.MaxValue, ErrorMessage = "O preço não pode ser negativo")]
-        public float Preco { get; set; }
+        [Range(0, double.MaxValue, ErrorMessage = "O preço não pode ser negativo")]
+        public decimal Preco { get; set; }
 
         public IngressoInteira() : base() { }
 
-        public IngressoInteira(float preco, int id, char fila, int numero, DateTime dataCompra,
+        public IngressoInteira(decimal preco, int id, char fila, int numero, DateTime dataCompra,
             Sessao? sessao = null, Cliente? cliente = null, Assento? assento = null)
             : base(id, fila, numero, dataCompra, sessao, cliente, assento)
         {
             Preco = preco;
         }
 
-        public override float CalcularPreco(float precoBase)
+        public override decimal CalcularPreco(decimal precoBase)
         {
             return Preco;
         }
