+++ cinecore/modelos/ItemPedidoAlimento.cs
@@ -15,8 +15,8 @@ namespace cinecore.modelos
         public int Quantidade { get; set; }
 
         [Required(ErrorMessage = "O preço é obrigatório")]
-        [Range(0, float.MaxValue, ErrorMessage = "O preço não pode ser negativo")]
-        public float Preco { get; set; }
+        [Range(0, double.MaxValue, ErrorMessage = "O preço não pode ser negativo")]
+        public decimal Preco { get; set; }
 
         public ProdutoAlimento? Produto { get; set; }
 
@@ -26,7 +26,7 @@ namespace cinecore.modelos
 
         public ItemPedidoAlimento() { }
 
-        public ItemPedidoAlimento(int id, int quantidade, float preco, ProdutoAlimento? produto = null)
+        public ItemPedidoAlimento(int id, int quantidade, decimal preco, ProdutoAlimento? produto = null)
         {
             Id = id;
             Quantidade = quantidade;
